\name{exactmaxsel-internal}
\alias{lower}
\alias{upper}
\alias{sumloop_cat}
\alias{subloop_cat}
\alias{loop_cat}
\alias{path}
\alias{underpmax}
\alias{loop_ord2}
\alias{subloop_ord2}


\title{Internal exactmaxsel Functions}
\description{
  Internal exactmaxsel functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
