\name{exactmaxsel-internal}
\alias{Points}
\alias{bipointok}
\alias{cc}
\alias{CC}
\alias{MatrixC}
\alias{loop}
\alias{lower}
\alias{upper}
\alias{sumloop}
\alias{subloop}
\alias{path}
\alias{underpmax}

\title{Internal exactmaxsel Functions}
\description{
  Internal exactmaxsel functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
